Revert "Store default config in job config and instance config in task config"
Summary:
This reverts commit 63754d816c71168da2d1f4f9dc099a7d29b47c4e.
Causing job get and job create APIs to run much more slowly and runtimes
have improved by 10% only. Will redo this diff after more discussion.

Reviewers: mu, #peloton

Reviewed By: mu, #peloton

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1390275

+++ jobmgr/tracked/task.go
@@ -218,7 +218,8 @@ func (t *task) getPostPreemptAction(ctx context.Context) (TaskAction, error) {
 func (t *task) getTaskPreemptionPolicy(ctx context.Context, jobID *peloton.JobID,
 	instanceID uint32, configVersion uint64) (*pb_task.PreemptionPolicy,
 	error) {
-	config, err := t.job.m.taskStore.GetTaskConfig(ctx, jobID, instanceID, configVersion)
+	config, err := t.job.m.taskStore.GetTaskConfig(ctx, jobID, instanceID,
+		int64(configVersion))
 	if err != nil {
 		return nil, err
 	}
